+++ packages/fhir_r6_bulk/lib/fhir_r6_bulk.dart
@@ -1,3 +1,5 @@
 export 'src/bulk_export.dart';
 export 'src/bulk_import.dart';
+export 'src/bulk_models.dart';
 export 'src/fhir_r6_bulk.dart';
+export 'src/ndjson_stream.dart';